pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uemura-jin-ai-agent"
version = "0.1.0"
description = "AI Agent to replace/assist Uemura Jin (上村仁)"
authors = [
    {name = "AI Development Team", email = "dev@sokagakuen.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "python-dateutil>=2.8.0",
    "mecab-python3>=1.0.0",
    "sudachipy>=0.6.0",
    "sudachidict-core>=20230110",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
uemura-ai = "uemura_ai.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["uemura_ai*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true